feat: create module structure for `module-manager` folder
