+++ packages/module-manager/lib/base/index.ts
@@ -1,0 +1,2 @@
+export * from './cli-args-processor.js';
+export * from './run-module-manager.js';
