+++ packages/module-manager/lib/base/run-module-manager.ts
@@ -1,15 +1,14 @@
-import { MANAGER_ACTION_PROMPT_CONFIG } from './const.js';
-import { deleteFeature } from './delete-feature.js';
-import { recoverFeature } from './recover-feature.js';
-import { promptSelection } from './utils.js';
-import manifest from '../../../chrome-extension/manifest.js';
+import manifest from '../../../../chrome-extension/manifest.js';
+import { MANAGER_ACTION_PROMPT_CONFIG } from '../const.js';
+import { promptSelection } from '../helpers/utils.js';
+import { deleteFeature, recoverFeature } from '../processing/index.js';
 import { execSync } from 'node:child_process';
 import { readFileSync, writeFileSync } from 'node:fs';
 import { resolve } from 'node:path';
-import type { CliActionType, ModuleNameType } from './types.js';
+import type { CliActionType, ModuleNameType } from '../types.js';
 import type { ManifestType } from '@extension/shared';
 
-const manifestPath = resolve(import.meta.dirname, '..', '..', '..', 'chrome-extension', 'manifest.ts');
+const manifestPath = resolve(import.meta.dirname, '..', '..', '..', '..', 'chrome-extension', 'manifest.ts');
 
 const manifestObject = JSON.parse(JSON.stringify(manifest)) as ManifestType;
 const manifestString = readFileSync(manifestPath, 'utf-8');
