+++ packages/module-manager/lib/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './utils.js';
+export * from './zip-utils.js';
