+++ packages/module-manager/lib/processing/delete-feature.ts
@@ -1,12 +1,12 @@
-import { DEFAULT_CHOICES, DELETE_CHOICE_QUESTION } from './const.js';
 import { deleteModule } from './modules-handler.js';
-import { processSelection } from './utils.js';
+import { DEFAULT_CHOICES, DELETE_CHOICE_QUESTION } from '../const.js';
+import { processSelection } from '../helpers/utils.js';
 import { existsSync, readdirSync } from 'node:fs';
 import { resolve } from 'node:path';
-import type { ChoicesType, ModuleNameType } from './types.ts';
+import type { ChoicesType, ModuleNameType } from '../types.ts';
 import type { ManifestType } from '@extension/shared';
 
-const pagesPath = resolve(import.meta.dirname, '..', '..', '..', 'pages');
+const pagesPath = resolve(import.meta.dirname, '..', '..', '..', '..', 'pages');
 const testsPath = resolve(pagesPath, '..', 'tests');
 
 const pageFolders = readdirSync(pagesPath);
@@ -15,8 +15,6 @@ export const deleteFeature = async (manifestObject: ManifestType, moduleName?: M
   const choices: ChoicesType = DEFAULT_CHOICES.filter(choice => {
     if (choice.value === 'background') {
       return !!manifestObject.background;
-    } else if (choice.value === 'tests') {
-      return existsSync(testsPath);
     }
 
     return pageFolders.includes(choice.value);
