+++ packages/module-manager/lib/processing/index.ts
@@ -1,0 +1,3 @@
+export * from './delete-feature.js';
+export * from './recover-feature.js';
+export * from './modules-handler.js';
